requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
everapi==0.1.1
freecurrencyapi==0.1.0
idna==3.10
Markdown==3.7
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
